feat: Separate the face recognition module.
